requirements.txt
# Existing dependencies
exa_py>=1.0.0
python-dotenv>=1.0.0

# LangChain core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.20

# LangGraph for stateful workflow orchestration
langgraph>=0.2.0

# OpenAI
openai>=1.0.0

# Google Gemini for image generation
google-generativeai>=0.3.0

# Supabase for storage and database
supabase>=2.0.0

# Mailchimp API
mailchimp3>=3.0.0

# Async HTTP for concurrent operations
aiohttp>=3.9.0

# Vector similarity (for learning system)
numpy>=1.24.0

# Optional: sentence embeddings for semantic search
sentence-transformers>=2.2.0

# Utilities
httpx>=0.25.0
tenacity>=8.2.0
